Fixed issue when saving links
* Now when saving each link a call to Reddit API is performed instead of
parsing the values in the HTML structure. This makes the saving a little
bit slower but more reliable, furthermore it allows me to create the
links upon page creation. Before I had to wait for the page to load and
the buttons were quite annoing because they were popping up at a
different time.

* Now a subreddit of the day is always stored, before the user had to
click on the selector to actually store the choice on disk.